Rename tty.py to pty_util.py
